aws/AWS-CDK/APIGateway/using-lambda-authorizers-on-routes/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.23.0
aws-cdk.aws-apigatewayv2-alpha==2.23.0a0
aws-cdk.aws-apigatewayv2-authorizers-alpha==2.23.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.23.0a0
aws-cdk.aws-lambda-python-alpha==2.23.0a0
black==22.3.0
boto3==1.22.11
botocore==1.25.11
cattrs==22.1.0
click==8.1.3
constructs==10.1.3
exceptiongroup==1.0.0rc5
flake8==4.0.1
jmespath==1.0.0
jsii==1.58.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
publication==0.0.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
